Revert "Add further polyfills for IE9 support"
This reverts commit ca68c26477681a6bde52ddf5357fc57cc85d0f62.

The polyfills helped to prevent all exceptions, but the positioning
of the tippy is not correct. This would require a further workaround
which is not justified in view of the particularity and low usage stat
of this browser (cf. https://github.com/FezVrasta/popper.js/issues/327).

The test for the creation of links is still helpful. If more browsers
have issues with tippy/popper, a fallback could be provided.